requirements.txt
semantic-kernel 
mcp[cli]
python-dotenv 
openai
psycopg2-binary
flask
requests